Unsubscribe from characteristics after disconnection
If a client subscribed to characteristics, disconnected and reconnected,
the old subscriptions stayed registered. If a new subscription to
notifications were added, the data was received twice on every callback.

This patch makes sure that the BleakClient keeps track of the active
subscriptions and explicitly removes them when a disconnect is called as
part of its cleanup.