feat(client): Add a method to get the exchanged MTU
BlueZ does not provide an easy way to access the MTU which forces us
to assume that it is 23 bytes in order to split our messages into
frames with a leading fragment byte.

By calling AcquireWrite [1], we can lock the current write and get a
writable socket and the initially exchanged MTU value with the central
in order to perform a Write Command (which we won't do as we just want
to get the MTU value and close the socket).

Note that BlueZ 5.50 relies only on AcquireWrite anyway.

[1]
https://git.kernel.org/pub/scm/bluetooth/bluez.git/tree/doc/gatt-api.txt#n115