feat(bluez): Wait for BlueZ to resolve services even when skipping discovery
Somehow BlueZ goes crazy if we do not wait for the services to be resolved
after a connec. The discovery becomes unavailable and a exception will be
thrown at any moment when ones try to do a DBus call.

We skip the complete services discovery but still wait for BlueZ to set
the `ServicesResolved` field to true.

This commit fixes the issue but unfortunately it also seems to increase the
disconnect time from around one second to three seconds, maybe BlueZ is
cleaning more things in the background?