fix(scanner): handle paired devices reconnecting
When a BLE device using privacy features is paired and bonded,
and subsequently tries to reconnect, BlueZ may not populate the UUIDs
in the advertisement data. This can cause the scanner to ignore the device.
The known / trusted device should be allowed to reconnect even if the UUIDs are not listed.